varnan/requirements.txt
streamlit
python-dotenv
PyMuPDF
crewAI
crewai-tools
pydub
edge-tts
litellm
pydantic
agentops
